Rename plugin to nvim-lsp-clangd-highlight
This plugin only deals with highlighting.

When I was looking for replacement to vim-lsp-cxx-highlight plugin I
skipped this plugin few times as I thought it's replacement config to
nvim-lspconfig.

This naming will be more clear about intentions of this plugin and make
it easier to find for people that want to find replacement for vim-lsp-cxx-highlight
that works with built in LSP